Phase 4a (#268): int8 cache storage + GPU-side widening cast
BLOSUM62 values fit tightly in [-4, +6] — they're actual integers.
The upstream encoder (``variable_length_to_fixed_length_vector_encoding``)
returns int8 natively. My Phase 1 encoding cache was casting int8 →
float32 before storage, throwing away a free 4× size reduction.

Changes:
  - ``EncodingCache._build``: preserve the encoder's native dtype
    rather than pin to float32. For BLOSUM62 that's int8 → memmap
    size shrinks 4× (1.8 GB → 450 MB for 517K peptides).
  - ``Class1NeuralNetwork.fit``'s inner batch loop: swap
    ``batch.float().to(device)`` → ``batch.to(device).float()`` so the
    int8→float32 widening happens on GPU, not CPU. Two wins:
      - H2D transfer is 4× smaller (int8 vs float32 over PCIe)
      - Cast on GPU is essentially free (wide-warp widening op,
        often fused into the first Linear's matmul downstream)
  - ``Class1NeuralNetwork.fit_generator``: same GPU-cast swap on the
    per-step H2D.

Values are preserved byte-for-byte (int8 → float32 widening is
lossless for integer inputs). All 91 Phase 1-4 tests pass, including
the bit-identical training integration gates.

Updated test_cached_matches_direct_byte_for_byte to reflect the new
contract: native-dtype storage, value-identical values. Added an
explicit guard that BLOSUM62 caches store int8, so a future regression
to fp32 storage (4× bigger, 2× slower H2D) fires loudly.

Empirical impact on pan-allele training:
  - Train cache for 517K peptides: 1.76 GB → 440 MB on disk
  - Pretrain cache for 256K peptides: ~900 MB → ~225 MB on disk
  - H2D bandwidth per batch: 4× reduction for peptide path

+++ mhcflurry/class1_neural_network.py
@@ -1861,9 +1861,13 @@ class Class1NeuralNetwork(object):
                 except StopIteration:
                     break
 
-                # Convert to tensors
-                peptide_tensor = torch.from_numpy(x_dict["peptide"]).float().to(device)
-                allele_tensor = torch.from_numpy(x_dict["allele"]).float().to(device)
+                # GPU-side widening cast (see Phase 4a #268): when the
+                # encoding cache stores int8 (BLOSUM62 native width),
+                # H2D transfer is 4× smaller; widening to fp32 on GPU
+                # is essentially free. For fp32-native encodings this
+                # is a no-op.
+                peptide_tensor = torch.from_numpy(x_dict["peptide"]).to(device).float()
+                allele_tensor = torch.from_numpy(x_dict["allele"]).to(device).float()
                 y_tensor = torch.from_numpy(y.astype(numpy.float32)).to(device)
 
                 optimizer.zero_grad()
@@ -2316,26 +2320,34 @@ class Class1NeuralNetwork(object):
             )
 
             for batch in loader:
-                peptide_batch = batch["peptide"].float().to(
+                # ``.to(device).float()`` (GPU-side widening cast) is
+                # strictly better than ``.float().to(device)``:
+                #   - When the cache stores int8 (BLOSUM62 native width),
+                #     the H2D transfer is 4× smaller.
+                #   - The widening cast on GPU is ~free (wide-warp op,
+                #     often fused into the first Linear's matmul).
+                #   - When the cache already stores fp32 (other
+                #     encodings), the cast is a no-op.
+                peptide_batch = batch["peptide"].to(
                     device, non_blocking=non_blocking_h2d
-                )
+                ).float()
                 y_batch = batch["y"].to(
                     device, non_blocking=non_blocking_h2d
                 )
 
                 inputs = {"peptide": peptide_batch}
                 if "allele" in batch:
-                    inputs["allele"] = batch["allele"].float().to(
+                    inputs["allele"] = batch["allele"].to(
                         device, non_blocking=non_blocking_h2d
-                    )
+                    ).float()
 
                 optimizer.zero_grad()
                 predictions = network(inputs)
                 weights_batch = None
                 if "weight" in batch:
-                    weights_batch = batch["weight"].float().to(
+                    weights_batch = batch["weight"].to(
                         device, non_blocking=non_blocking_h2d
-                    )
+                    ).float()
                 loss = loss_obj(predictions, y_batch, sample_weights=weights_batch)
                 regularization_penalty = self._regularization_penalty(
                     regularization_parameters,
